+++ Tacta.EventStore/Repository/StoredEvent.cs
@@ -7,7 +7,7 @@ namespace Tacta.EventStore.Repository
         public string AggregateId { get; set; }
         public string Aggregate { get; set; }
         public int Version { get; set; }
-        public int Sequence { get; set; }
+        public long Sequence { get; set; }
         public Guid Id { get; set; }
         public string Name { get; set; }
         public DateTime CreatedAt { get; set; }
